requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Helmet==0.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
bandit==1.7.5
safety==2.3.5
requests==2.31.0 